requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-Session==0.5.0
Flask-Caching==2.0.2
mysql-connector-python==8.1.0
python-dotenv==1.0.0
PyMySQL==1.1.0
twilio==8.5.0
Werkzeug==2.3.7
gunicorn==21.2.0
